♻️ Refactor: 코드 복잡성에 따른 헤더 컴포넌트 분리
