+++ src/components/ModalHeader/styles.module.scss
@@ -1,0 +1,26 @@
+$titleColor: #f7f7f8;
+
+@mixin font-color($color, $weight, $size, $lineHeight) {
+  font: {
+    family: Pretendard;
+    weight: $weight;
+    size: $size;
+  }
+  line-height: $lineHeight;
+  color: $color;
+}
+
+.Container {
+  display: flex;
+  width: 100%;
+  justify-content: flex-end;
+  > h2 {
+    @include font-color($titleColor, 600, 18px, 21.48px);
+  }
+  > .HeaderCloseButton {
+    cursor: pointer;
+  }
+  &.titleActive {
+    justify-content: space-between;
+  }
+}
