+++ src/components/ModalMobileHeader/styles.module.scss
@@ -1,0 +1,29 @@
+$titleColor: #f7f7f8;
+
+@mixin font-color($color, $weight, $size, $lineHeight) {
+  font: {
+    family: Pretendard;
+    weight: $weight;
+    size: $size;
+  }
+  line-height: $lineHeight;
+  color: $color;
+}
+
+.Container {
+  display: flex;
+  width: 100%;
+  justify-content: space-between;
+  flex-direction: row-reverse;
+  > h2 {
+    @include font-color($titleColor, 600, 18px, 21.48px);
+    display: flex;
+    flex-grow: 3;
+    justify-content: center;
+  }
+  > .BackButton {
+    flex-grow: 0;
+    justify-content: flex-start;
+    cursor: pointer;
+  }
+}
