+++ src/pages/TestPage/index.jsx
@@ -1,6 +1,8 @@
 import Sample from '@/components/Sample/Sample';
 import useModal from '@/hooks/useModal';
-import Modal from '@/Modal';
+import Modal from '@/components/Modal';
+import ModalHeader from '@/components/ModalHeader';
+import ModalMobileHeader from '@/components/ModalMobileHeader';
 
 const TestPage = () => {
   const [isOpen, openModal, closeModal] = useModal();
@@ -11,7 +13,9 @@ const TestPage = () => {
       <Sample />
 
       <button onClick={openModal}>Open Modal</button>
-      <Modal isOpen={isOpen} title="모달" onClose={closeModal} isFull="true">
+      <Modal isOpen={isOpen} title="모달" onClose={closeModal}>
+        {/* <ModalHeader title="모달" onClose={closeModal} /> */}
+        <ModalMobileHeader title="모달" onClose={closeModal} />
         <div style={{ color: '#fff' }}>바디입니다.</div>
       </Modal>
     </div>
